Struct gfx_gl::FnPtr

pub struct FnPtr {
    // some fields omitted
}

Methods

impl FnPtr

fn is_loaded(&self) -> bool

Returns true if the function has been successfully loaded.

If it returns false, calling the corresponding function will fail.